Provide available fields programmatically
Admins shouldn't have to write JSON schema just to
configure the fields here. A 'library' of fields that
are configurable should be provided, and admins can just
'pick' from these.

Pluggy with entrypoints is used to let arbitrary python
packages provide available fields from which admins can
select what is displayed to users.

Fields are now namespaced, but arbitrarily. We add a
z2jh namespace with some useful parameters - although that
should probably go into a kubespawner namespace?

The 'InterfaceSelector' is turned into a generic RadioWidget,
and we can use it for a bunch of other things too!

We store & ship the schema for the config along with the config,
since these two are intricately linked. The traitlet name is
defined in the schema now, so we need it anyway to figure
out what traitlets to set where.